layered_grpc/api_gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
grpcio==1.60.0
grpcio-tools==1.60.0
pydantic==2.9.2
protobuf==4.25.1
